refactored fibonacci and input to take 128 bit data as originally program would overflow since it was taking in and only computing 32 bit numbers

+++ src/main.rs
@@ -5,13 +5,13 @@ fn main() {
     println!("The fibonacci of given number {} is {}", number, calculate_fib(number));
 }
 
-fn input_num_to_calculate() -> i32 {
+fn input_num_to_calculate() -> i128 {
     loop {
         println!("Enter number to calculate fibonacci of:");
         let mut number = String::new();
         io::stdin().read_line(&mut number).expect("Failed to read user input!");
 
-        let number: i32 = match number.trim().parse() {
+        let number: i64 = match number.trim().parse() {
             Ok(num) => break num,
             Err(_) => {
                 println!("Please enter a valid number!");
@@ -21,7 +21,7 @@ fn input_num_to_calculate() -> i32 {
     }
 }
 
-fn calculate_fib(num: i32) -> i32 {
+fn calculate_fib(num: i128) -> i128 {
     let mut a = 0;
     let mut b = 1;
     let mut c = 0;
